+++ shell.go
@@ -10,7 +10,6 @@ import (
 	"bytes"
 	"fmt"
 	"io"
-	"log"
 	"os"
 	"time"
 
@@ -81,44 +80,6 @@ func (c *Connect) CmdShell(session *ssh.Session, command string) (err error) {
 	err = session.Wait()
 	if err != nil {
 		return
-	}
-
-	return
-}
-
-func (c *Connect) setupShell(session *ssh.Session) (err error) {
-	// set FD
-	session.Stdin = os.Stdin
-	session.Stdout = os.Stdout
-	session.Stderr = os.Stderr
-
-	// Logging
-	if c.logging {
-		err = c.logger(session)
-		if err != nil {
-			log.Println(err)
-		}
-	}
-	err = nil
-
-	// Request tty
-	err = RequestTty(session)
-	if err != nil {
-		return err
-	}
-
-	// x11 forwarding
-	if c.ForwardX11 {
-		err = c.X11Forward(session)
-		if err != nil {
-			log.Println(err)
-		}
-	}
-	err = nil
-
-	// ssh agent forwarding
-	if c.ForwardAgent {
-		c.ForwardSshAgent(session)
 	}
 
 	return
